requirements.txt
asgiref==3.11.1
attrs==25.4.0
colorama==0.4.6
Django==6.0.3
django-extensions==4.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.15
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
